Updated error handling so when a Broker fails and a new leader is identified the Storm/Trident side doesn't crash.
However this is way less than ideal. Because it isn't clear what will happen if we change the data in the Partition object from Storm we are playing games under the covers to map the current Broker to data in the Partition object.

Need help from Nathan to understand the impact.